feat: find archetype for components

+++ common/src/ecs/entity.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Entity {
-    pub archetype_id: u32,
+    pub archetype_id: u64,
     pub row: u32,
 }
